zinoma::config::yaml
Module schema

Structs

Dependencies

List of targets that must complete successfully before this target can be built.

InputResources

List of artifacts that this target depends on.

OutputResources

List of artifacts produced by this target.

Project

Schema of the build flow configuration file zinoma.yml.

Enums

InputResource

OutputResource

Target

A target is a command or a set of commands to run as part of your build flow.