Struct zinoma::config::yaml::schema::Target

pub struct Target {
    pub dependencies: Vec<String>,
    pub build: Option<String>,
    pub input: Vec<Resource>,
    pub output: Vec<Resource>,
    pub service: Option<String>,
}

Fields
dependencies: Vec<String>

Targets that must complete successfully before the target can run

It should be an array of strings.

If a target fails, targets that depend on it will not be executed.

Example

targets:
  target1:
  target2:
    dependencies: [target1]
  target3:
    dependencies: [target2]

In this example, target1 must complete successfully before target2 begins, while target3 waits for target2 to complete.

zinoma target2 will run sequentially target1 and target2.

zinoma target3 will run sequentially target1, target2 and target3.

build: Option<String>

Build script of the target

It should be a string which can have one or multiple lines.

Example

targets:
  create_my_file:
    build: |
      mkdir -p deep/dir
      touch deep/dir/my_file

In this example, running zinoma create_my_file will execute the commands mkdir -p deep/dir and touch deep/dir/my_file sequentially.

input: Vec<Resource>

Resources the target depends on

input should be an array of objects.

Specifying a target's input enables the incremental build for this target. This means that, at the time of executing the target, Žinoma will skip its build if its input resources (and output resources, if any) have not changed since its last successful completion.

Example

targets:
  npm_install:
    input:
      - paths: [package.json, package-lock.json]
    build: npm install

In this example, running zinoma npm_install once will execute npm install. Subsequent runs of zinoma npm_install will return immediately — until the content of package.json or package-lock.json is modified.

output: Vec<Resource>

Resources produced by the target

This keyword lists resources identifying the artifacts produced by this target.

Similarly to input, it should be an array of objects.

If the --clean flag is provided to zinoma, the files or directories specified in output.paths will be deleted before running the build flow.

The incremental build takes in account the target output. Just like with input, if any of the target output resources were altered since its previous successful execution, the target state will be invalidated and its build will be run again.

Example

targets:
  npm_install:
    input:
      - paths: [package.json, package-lock.json]
    output:
      - paths: [node_modules]
    build: npm install

In this example, running zinoma npm_install will return immediately in case package.json, package-lock.json and node_modules were not modified since the last completion of the target.

Running zinoma --clean npm_install will start by deleting node_modules, then will run npm install.

service: Option<String>

Script starting a long-lasting process to execute upon build success

Specifies a command to run upon successful build of the target.

It should be a string.

This keyword is meant to enable the execution of long-lasting commands, such as servers.

If the targets to run do not define services, zinoma will automatically exit after all builds ran to completion. On the contrary, if at least one target defines a service, zinoma will keep running even after all builds completed, so that the services can remain alive.

In watch mode (when the --watch flag is passed to zinoma), services are restarted every time the target build runs to completion.

Example

targets:
  npm_server:
    input:
      - paths: [package.json, index.js]
    build: npm install
    service: npm start

In this example, zinoma npm_server will run npm install and then npm start.