Enum zinoma::engine::target_actor::ActorInputMessage

pub enum ActorInputMessage {
    Requested {
        kind: ExecutionKind,
        requester: ActorId,
    },
    Unrequested {
        kind: ExecutionKind,
        requester: ActorId,
    },
    Ok {
        kind: ExecutionKind,
        target_id: TargetId,
        actual: bool,
    },
    Invalidated {
        kind: ExecutionKind,
        target_id: TargetId,
    },
}

Variants

Requested

Indicates the execution of the build scripts or services behind this target are requested.

This message should only be sent to direct dependencies.

Fields of Requested

kind: ExecutionKindrequester: ActorId

Unrequested

Indicates the execution of the build scripts or services behind this target are no more requested by the provided requester.

This message should only be sent to direct dependencies.

Fields of Unrequested

kind: ExecutionKindrequester: ActorId

Ok

Indicates the execution of the build scripts behind the provided target are Ok.

If kind is Build, Ok means one of the following:

There is no build script behind this target;

All build scripts have been executed or skipped, and therefore, their output resources are available.

If kind is Service, Ok means one of the following:

There is no service behind this target;

All services have been started and are currently running.

actual should be true if there is an actual build/service behind this target.

This message should only be sent to build requesters.

Fields of Ok

kind: ExecutionKindtarget_id: TargetIdactual: bool

Invalidated

Indicates the build scripts or services behind the target are not OK anymore.

The requester should invalidate the previously sent Ok.

Fields of Invalidated

kind: ExecutionKindtarget_id: TargetId